chore: organise test directory better
